Split long e2e specs
